% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGGMSampler.R
\name{BGGMSampler}
\alias{BGGMSampler}
\title{Bayesian Gaussian Graphical Model Sampler}
\usage{
BGGMSampler(
  n,
  S,
  C,
  beta = 0.5,
  bPrior = 3,
  DPrior = diag(ncol(C)),
  burnin = 0,
  nmcmc = 1
)
}
\arguments{
\item{n}{Number of observations.}

\item{S}{Sample Covariance  multiplied by n.}

\item{C}{Initial Precision Matrix for the MCMC.}

\item{beta}{Prior probability for each edge.}

\item{bPrior}{Prior dedrees of freedom. Defaults to 3.}

\item{DPrior}{Prior location SPD Matrix. Defaults to the identity matrix.}

\item{burnin}{Number of sample to burn in in the MCMC.}

\item{nmcmc}{Number of MCMC samples desired as output. That is without
considering the burn-in period.}
}
\value{
List containg two arrays. One for the Precison matrices and another
  one for the adjacency matrices.
  \describe{
    \item{samC}{An Array of Precicion matrices, in which the last dimesion
      goes through every Precision matrix.}
    \item{samE}{An Array of Adjacency matrices, in which the last dimesion
      goes through every adjacency matrix.}
 }
}
\description{
This function generates nmcmc MCMC samples for a Bayesian Gaussian Graphical
Model following the exchange method proposed by Wang and Li, in Efficient
Gaussian graphical Model determination under G-Wishart prior distributions.
}
\author{
Rene Gutierrez Marquez
}
