% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTGM.R
\name{BTGM}
\alias{BTGM}
\title{Performs Bayesian Tensor Graphical Model Estimation}
\usage{
BTGM(t, b = NULL, D = NULL, C = NULL, beta = NULL, burnin = 0, nmcmc = 1)
}
\arguments{
\item{t}{List of Sample Tensors}

\item{b}{A vector with the Prior degrees of Freedom for each precision
matrix. If NULL it sets the prior for each presicion at 3.}

\item{D}{List of Prior Scale matrices for each precision matrix. If NULL
it defaults to the corresponding Identity matrices for each presicion
matrix.}

\item{C}{List of Initial Values of the Precicion Matrices to start the
MC markov chain. If NULL each presicion matrix is initialized as the
corresponding identity matrix.}

\item{beta}{A vector containing the prior probability of having and edge
between to vertices for each precision. If NULL it sets the value at 0.5
for every edge matrix.}

\item{burnin}{Number of sample to burn in in the MCMC.}

\item{nmcmc}{Number of MCMC samples desired as output. That is without
considering the burn-in period.}
}
\value{
List containg two other lists. One for the Precison matrices and
  another one for the adjacency matrices.
\describe{
  \item{samC}{A list of arrays for Precicion matrices, the list goes through
  every precision matrix.}
  \item{samE}{A list of arrays for Adjacency matrices, the list goes through
  every adjacency matrix.}
}
}
\description{
This function generates nmcmc MCMC samples for a Bayesian Tensor Graphical
Model. Each precision matrix is modeled according to a GWishart distribution.
The output, returns the samples for each presion matrix as well as the
adjacency matrix samples.
}
\author{
Rene Gutierrez Marquez
}
