% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datPos.R
\name{datPos}
\alias{datPos}
\title{Computes Graphic outputs of the Marginal Posterior of the Precision Matrices}
\usage{
datPos(posPre, reaPre = NULL, sel)
}
\arguments{
\item{posPre}{A list of arrays for Precicion matrices, the list goes through
every precision matrix.}

\item{reaPre}{A list of the real precision matrices. By default it is set to
NULL, in which case no real values are ploted for reference.}

\item{sel}{A list of selected entries for each Matrix}
}
\value{
A data frame used to plot the trace plot and density of each entry.
 It specifies the matrix is comming from and the entry.
}
\description{
Computes the trace plot and density plot of one entry of a Presicion Matrix.
}
\author{
Rene Gutierrez Marquez
}
