% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgMat.R
\name{edgMat}
\alias{edgMat}
\title{Generates a Random Adjacency or Edge Matrix}
\usage{
edgMat(matSiz, spa, upp = TRUE)
}
\arguments{
\item{matSiz}{A scalar specifying the size of the matrix. As the result is
a square matrix only a scalar is requiered.}

\item{spa}{The degree of sparsity specified with a number from 0 to 1.}

\item{upp}{An indicator indicating if the result should be an Upper
Triangular Matrix. The default is TRUE.}
}
\value{
A Random Square Matrix encoding the Adjacency of vertices on a
        graph. If upp is TRUE (defualt) the result would by an upper
        triangular matrix.
}
\description{
This function generates Random Adjacency or Edge Matrix of a specific size
and degree of sparsity.
}
\author{
Rene Gutierrez Marquez
}
