% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entPos.R
\name{entPos}
\alias{entPos}
\title{Computes Graphic outputs of the Marginal Posterior of the Precision Matrices}
\usage{
entPos(posPre, matNum = 1, i = 1, j = 2)
}
\arguments{
\item{posPre}{A list of arrays for Precicion matrices, the list goes through
every precision matrix.}

\item{i}{Row number of the entry.}

\item{j}{Column number of the entry}
}
\value{
A data frame used to plot the trace plot and density of an entry. It
  specifies the matrix is comming from and the entry. Test
}
\description{
Computes the trace plot and density plot of one entry of a Presicion Matrix.
}
\author{
Rene Gutierrez Marquez
}
