% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatMat.R
\name{heatMat}
\alias{heatMat}
\title{Creates a Heatmap Database given a Matrix}
\usage{
heatMat(ll, modelNames = NULL)
}
\arguments{
\item{ll}{A list of lists. The outer list is indexed by model. The
inner list is indexed by matrix.}

\item{modelNames}{The names of the models to use in the data Frame.}
}
\value{
A data Frame to create heatmaps in ggplot. With columns indicating
  the Matrix and Model
}
\description{
This function creates a data frame to create Heatmaps for each model and for
each Matrix
}
\author{
Rene Gutierrez Marquez
}
