% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kModMat.R
\name{kModMat}
\alias{kModMat}
\title{k-mode Tesnor Matricization}
\usage{
kModMat(tensor, mode = 1)
}
\arguments{
\item{tensor}{An array for the Tensor to matricize.}

\item{k}{An integer specifying the mode to matricize.}
}
\value{
A matrix that is the k-mode matricization of the tensor provided.
}
\description{
This function computes the k-mode Tesnor Matricization of a tensor.
}
\author{
Rene Gutierrez Marquez
}
