% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lMVGamma.R
\name{lMVGamma}
\alias{lMVGamma}
\title{Logarithm of the Multivariate Gamma Function}
\usage{
lMVGamma(p = 1, a = 1)
}
\arguments{
\item{p}{Dimension parameter.}

\item{a}{Real number to evaluate the function.}
}
\value{
The Logarithm of the Multivariate Gamma Function.
}
\description{
This function computes the Multivariate Gamma Function.
}
\author{
Rene Gutierrez Marquez
}
