% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logConsW.R
\name{logConsW}
\alias{logConsW}
\title{Logarithm of the Normalizing Constant of the Wishart Distribution}
\usage{
logConsW(v, S)
}
\arguments{
\item{v}{Degrees of freedom.}

\item{S}{Scale matrix parameter.}
}
\value{
The Logarithm of the Normalizing Constant of the Wishart
  Distribution
}
\description{
This function computes the Logarithm of the Normalizing Constant of the
Wishart Distribution
}
\author{
Rene Gutierrez Marquez
}
