% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logH.R
\name{logH}
\alias{logH}
\title{Logarithm of H}
\usage{
logH(n, S, C, bPrior = 100, DPrior = diag(ncol(C)), i = 1, j = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{S}{Sample Covariance multiplied by n.}

\item{C}{Initial Precision Matrix for the MCMC.}

\item{bPrior}{Prior dedrees of freedom. Defaults to 3.}

\item{DPrior}{Prior location SPD Matrix. Defaults to the identity matrix.}

\item{i}{Row index.}

\item{j}{Column index.}
}
\value{
Logarithm of the H function of the Acceptance Rate.
}
\description{
This function computes the function H that appears in the Metropolis Odds as
propose in Wang and Li, in Efficient Gaussian graphical Model determination
under G-Wishart prior distributions.
}
\author{
Rene Gutierrez Marquez
}
