% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logJ.R
\name{logJ}
\alias{logJ}
\title{Logarithm of J}
\usage{
logJ(h, B, a11)
}
\arguments{
\item{h}{Degrees of Freedom for the Wishart Distribution}

\item{B}{2 by 2 sub-matrix of the precision matrix.}

\item{a11}{Entry 1,1 from the A matrix.}
}
\value{
The Logarithm of J.
}
\description{
This function computes the function J that appears in the Metropolis Odds as
propose in Wang and Li, in Efficient Gaussian graphical Model determination
under G-Wishart prior distributions.
}
\author{
Rene Gutierrez Marquez
}
