% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkronecker.R
\name{mkronecker}
\alias{mkronecker}
\title{Perfroms the Kronecker Product from a list of Matrices}
\usage{
mkronecker(C)
}
\arguments{
\item{C}{The list of matrices on which the Kronecker product is to be
performed}
}
\value{
A matrix which is the kronecker product of the list of matrices
        with number of rows eaul to the product of the number of
        rows of every matrix, and number of columns equal to the number of
        columns of every matrix in the list.
}
\description{
This function performs the kronecker product of a list of matrices in the
order in which they appear on the list with the first element of the list
beign the rightmost matrix in the product and the last element of the list
beign the leftmost matrix in the product.
}
\author{
Rene Gutierrez Marquez
}
