% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modComAdj.R
\name{modComAdj}
\alias{modComAdj}
\title{Computes Several performance statistics for the Tensor Gaussian Graphical
Model}
\usage{
modComAdj(modelList, trueModel)
}
\arguments{
\item{modelList}{A list of lists. The outer list is indexed by model. The
inner list is indexed by Matrix.}

\item{trueModel}{A list with the true Matrices}
}
\value{
A list with different performance statistics in comparison with the
  true model.
\describe{
  \item{acc}{Accuracy for each model.}
  \item{tpr}{True positive Rate for each model.}
  \item{tnr}{True Negative Rate for each model.}
}
}
\description{
This function computes the following statistics for each model provided.
\itemize{
  \item Accuracy.
  \item True Positive Rate.
  \item True Negative Rate.
}
}
\author{
Rene Gutierrez Marquez
}
