% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modComLik.R
\name{modComLik}
\alias{modComLik}
\title{Computes the log-likelihood of the Tesor Normal for each model estimate.}
\usage{
modComLik(modelList, tensors)
}
\arguments{
\item{modelList}{A list of lists. The outer list is indexed by model. The
inner list is indexed by Matrix.}

\item{tensors}{A list with the sample of tensors}
}
\value{
A vector with the log-likelihood for each model.
}
\description{
Computes the log-likelihood of the Tesor Normal for each model estimate
given a sample of tensors.
}
\author{
Rene Gutierrez Marquez
}
