% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modComPre.R
\name{modComPre}
\alias{modComPre}
\title{Computes Several performance statistics for the Tensor Gaussian Graphical
Model}
\usage{
modComPre(modelList, trueModel, large = TRUE)
}
\arguments{
\item{modelList}{A list of lists. The outer list is indexed by model. The
inner list is indexed by Matrix.}

\item{trueModel}{A list with the true Matrices.}

\item{large}{A boolean indicating if the presicion matrices are large,
to avoid computing the kronacker norm.}
}
\value{
A list with different performance statistics in comparison with the
  true model.
\describe{
  \item{kroFro}{Frobenious Norm of the Kronecker product for each model.}
  \item{kroInf}{Infinity Norm of the Kronecker product for each model.}
  \item{sinFro}{Average Frobenious Norm of each matrix for each model.}
  \item{sinInf}{Average Infinity Norm of each matrix for each model.}
}
}
\description{
This function computes the following statistics for each model provided.
\itemize{
  \item Frobenious Norm of the Kronecker product.
  \item Infinity Norm of the Kronecker product.
  \item Average Frobenious Norm of each matrix.
  \item Average Infinity Norm of each matrix.
}
}
\author{
Rene Gutierrez Marquez
}
