% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorConRatio.R
\name{priorConRatio}
\alias{priorConRatio}
\title{Logarithm of the Prior Constant Ratio Approximation}
\usage{
priorConRatio(C, bPrior = 100, DPrior = diag(ncol(C)), i = 1, j = 1)
}
\arguments{
\item{C}{Initial Precision Matrix for the MCMC.}

\item{bPrior}{Prior dedrees of freedom. Defaults to 3.}

\item{DPrior}{Prior location SPD Matrix. Defaults to the identity matrix.}

\item{i}{Row index.}

\item{j}{Column index.}
}
\value{
Logarithm of the Prior Constant Ratio Approximation.
}
\description{
This function computes the Prior Constant Ratio Approximation that appears in
the Metropolis Odds as propose in Wang and Li, in Efficient Gaussian
graphical Model determination under G-Wishart prior distributions.
}
\author{
Rene Gutierrez Marquez
}
