% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sK.R
\name{Sk}
\alias{Sk}
\title{Computes the Conditional Statistic S_k}
\usage{
Sk(tensors, precisions, k)
}
\arguments{
\item{tensors}{List of Sample Tensors}

\item{precisions}{A list of the precision matrices}

\item{k}{An integer specifying the precision matrix for wich the S_k
statistic is desired.}
}
\value{
A sqaure matrix of the same dimension as the precision matrix k in
  the list with the Conditional Statistic S_k for the profile likelihood of
  the precision matrix C_k
}
\description{
This function computrd the Conditional Statistic S_k for the profile
likelihood of the precision matrix C_k of a Tensor Normal distribution. That
is the quadratic form of the k-mode matricization of the sample tensor with
the Kronecker product of the remaining Presision matrices.
}
\author{
Rene Gutierrez Marquez
}
