% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaPDM.R
\name{spaPDM}
\alias{spaPDM}
\title{Generates Sparse Symmetric Positive Definite Random Matrices.}
\usage{
spaPDM(b, E)
}
\arguments{
\item{b}{Degrees of Freedom for Sampling from the Gwishart distribution.}

\item{E}{The Edge or Adjacency matrix for sampling from the GWishart
distribution.}
}
\value{
A random Sparse Symmetric Positive Definite matrix that has zero
        entries whenever the adjacency matrix has a zero entry.
}
\description{
This function generates Sparse Symmetric Positive Definitive Matrices
according to the adjacency matrix provided and certain degrees of freedom.
The whole procedure follows, ...
}
\author{
Rene Gutierrez Marquez
}
