% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaPreMatGen.R
\name{spaPreMatGen}
\alias{spaPreMatGen}
\title{Creates Sparse Precision Matrices}
\usage{
spaPreMatGen(p = c(10, 10, 10), type = "R", sparsity = NULL)
}
\arguments{
\item{p}{A vector specifying the dimension size of each precision
graph, it defaults to a 3 dimesional Tensor with each dimension os size 10.}

\item{type}{A charachter strin specifying the matrix structure:
\describe{
  \item{"R"}{For a Random Matrix, as described in Wang & Li.}
  \item{"C"}{For a "Cyclic Graph" Matrix.}
}
It defaults to a random matrix "R".}

\item{sparsity}{A vector with the requiered sparsity for the Random Matrix.
If NULL it defaults to a 10\% sparsity level for each precision matrix.}
}
\value{
A list of lists containing the Presion Matrices, the Covariance
  Matrices and the Adjacency Matrices.Each list is as follows:
\describe{
\item{P}{A list of Precision Matrices.}
\item{S}{A list of Covariance Matrices.}
\item{E}{A list of Adjacency Matrices.}
}
}
\description{
This Funtion Creates Sparse Precision Matrices with a shape specified by the
user.
}
\author{
Rene Gutierrez Marquez
}
