% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLasso.R
\name{TLasso}
\alias{TLasso}
\title{Performs Tensor Lasso}
\usage{
TLasso(t, iter = 1, lambda = NULL, identy = "F", norm = FALSE, search = FALSE)
}
\arguments{
\item{t}{List of Sample Tensors}

\item{iter}{Number of iterations}

\item{lambda}{A vector to use as the penalties for each precision matrix.}

\item{norm}{A boolena indicating if the norm of the precision matrices
should remain constant through the iterations.}

\item{search}{A boolean indicating if a penalty search based on likelihood
should be performed.}

\item{identi}{A method to obtain identifiability.
'F' for Frobenius norm normalization.
'C' for correlation normalization.
'1' for dividing by the first entry of the precision matrix.}
}
\value{
List containg the precision matrices.
}
\description{
This function performs Tensor Lasso under different specifications.
}
\author{
Rene Gutierrez Marquez
}
