% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNormalSampler.R
\name{TNormalSampler}
\alias{TNormalSampler}
\title{Generates random samples from the Tensor Normal Distribution}
\usage{
TNormalSampler(n = 1, Tmean = NULL, SigmaList, listInd = TRUE)
}
\arguments{
\item{n}{Number of samples generated.}

\item{Tmean}{Tensor mean. Must be of the same dimensions as the
specified by the list of Covariance matrices provided.
If NULL (default) a zero tensor mean is generated.}

\item{SigmaList}{A list of the covariances matrices.}

\item{listInd}{A boolean specifying if the samples should be provided in
a list or an array. If TRUE (default) a list is provided.}
}
\value{
Either a list with the samples or an array whose last dimension is
        the index of the sample.
}
\description{
This function generates n samples from the Tensor Normal distribution as
specified by its covariances matrices.
}
\author{
Rene Gutierrez Marquez
}
