% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updij.R
\name{updij}
\alias{updij}
\title{Updates the Entries of the Precision Matrix}
\usage{
updij(C, bPost, DPost, i, j, proij)
}
\arguments{
\item{C}{Initial Precision Matrix for the MCMC.}

\item{i}{Row index.}

\item{j}{Column index.}

\item{proij}{Boolean indicatting an edgeif TRUE and no edge if FALSE.}

\item{bPrior}{Prior dedrees of freedom. Defaults to 3.}

\item{DPrior}{Prior location SPD Matrix. Defaults to the identity matrix.}
}
\value{
An updated Precision matrix.
}
\description{
This function updates the Entries (j,j), (i,j) and (j,i) of the Precision Matrix according to the proposed Edge
}
\author{
Rene Gutierrez Marquez
}
